backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "beautifulsoup4>=4.13.5",
    "copilotkit>=0.1.65",
    "email-validator>=2.3.0",
    "fastapi>=0.115.14",
    "google-generativeai>=0.8.5",
    "holidays>=0.81",
    "httpx>=0.28.1",
    "joblib>=1.5.2",
    "langchain-google-genai>=2.0.10",
    "lxml>=6.0.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[google]>=1.0.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.36.0",
]
